Check off which User stories I've satisfied: "As a user, I want to see how layouts (like QWERTY, Colemak, Dvorak, Workman) are rated against each other by letting them compete." Doable via Tournaments and default values for layout
"As a user, I want to create a keyboard layout and add it to a competition
 between other layouts" Doable through al command and tournaments

"As a user, I want to remove layouts that I'm not interested in judging, or layouts that aren't applicable to me (like Neo, which is optimized for typing German)." Doable through remove commands.

"As a user, I want to compute specific effort values for typing a certain corpora on a specific keyboard." Doable through tournaments.